Update everything, move away from ember-cli-typescript (#270)
* Remove deprecated `loc`

* v3.9.0...v3.28.0

* Lint & Prettier fix; remove unused generated test

* Update CI with new ember-try scenarios

* Update types and copy ts setup from ember-test-helpers

* Lint fixup

* `yarn test` runs ember:try, make not-explicitly-ember-try tests in CI
only run ember tests

* dont fail fast on matrix tests

Failing fast makes it difficult to diagnose the subset of the matrix
that is failing beyond the first job that fails

* Fix TypeScript compilation

Currently this is clearly suboptimal, once we learn more about TS in our projects. We can revisit ember-cli-typescript (or similar) to handle this for us automatically. Additionally, we plan embroider v2 to enable this “for free”

* Re-roll lock file

* Upgrade some more deps

* Fix gitignore + npmignore files

* updates

Co-authored-by: Stefan Penner <stefan.penner@gmail.com>